refactored with Token class
Now uses InfixTokeniser to get a list of Tokens. The method operates entirely on these Tokens by rearranging them into a postfix order. A list of the Tokens in the correct order is returned. This simplification means that the other private methods aren't needed and that the algorithm can be simplified, as the infix expression is tokenised by a separate class.